+++ python/datamodel/models/__init__.py
@@ -12,3 +12,6 @@
 
 
 from __future__ import print_function, division, absolute_import
+
+from .releases import releases
+from .surveys import surveys, phases